+++ gubbi_common/__init__.py
@@ -2,6 +2,6 @@
 
 from gubbi_common.constants import RESOURCE_DOCUMENTATION_URL
 
-__version__ = "0.15.0"
+__version__ = "0.16.0"
 
 __all__ = ["RESOURCE_DOCUMENTATION_URL"]